requirements.txt
streamlit==1.48.0
opencv-python==4.12.0.88
pillow==11.3.0
numpy==2.2.6
streamlit-image-comparison==0.0.4